+++ backend/src/Models/Shelves/SUpdate.cs
@@ -5,5 +5,5 @@ namespace Jupiter.Models.Shelves;
 public class SUpdate
 {
     public required int ShelfId { get; set; }
-    public required int?[] ProductIds { get; set; }
+    public required ulong?[] ProductIds { get; set; }
 }
