+++ backend/src/Models/Stock/SProduct.cs
@@ -1,6 +1,6 @@
 namespace Jupiter.Models.Stock;
 
 public class SProduct {
-    public required int ProductId { get; set; }
+    public required ulong ProductId { get; set; }
     public required int Amount { get; set; }
 }
